rt#27684 -- check if readonly mode is on
git-svn-id: svn+ssh://svn.wikia-code.com/svn/wikia/trunk@16300 e846bd91-c042-42a7-a218-4b65e2c1904d

+++ extensions/3rdparty/LyricWiki/server.php
@@ -2050,8 +2050,10 @@ function requestStarted($funcName, $requestData){
 // of the request.
 ////
 function requestFinished($id){
-	if(defined('TRACK_REQUEST_RUNTIMES') && TRACK_REQUEST_RUNTIMES){
-		$db = lw_connect();
-		mysql_query("DELETE FROM apiRequests WHERE id=$id", $db);
+	if( !wfReadOnly() ) {
+		if(defined('TRACK_REQUEST_RUNTIMES') && TRACK_REQUEST_RUNTIMES){
+			$db = lw_connect();
+			mysql_query("DELETE FROM apiRequests WHERE id=$id", $db);
+		}
 	}
 } // end requestFinished()
